requirements.txt
flet
pydirectinput_rgx